% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_poc_nonadap.R
\name{single_poc}
\alias{single_poc}
\title{Single Arm Calculator (Non-Adaptive)}
\usage{
single_poc(soc = 0.25, exp = 0.4, alpha = 0.05, pwr = 0.8)
}
\arguments{
\item{soc}{Standard of care efficacy (0 - 1 value)}

\item{exp}{Expected experimental treatment efficacy (0 - 1 value)}

\item{alpha}{Desired alpha level (0 - 1 value)}

\item{pwr}{Desired power (0 - 1 value)}
}
\value{
Displays the necessary amount of patients needed for the sample size to satisfy user defined parameters.
}
\description{
Single Arm Calculator (Non-Adaptive)
}
\examples{
library(ParamCalc)
single_poc()
single_poc(soc = 0.27, exp = 0.36, alpha = 0.03, pwr = 0.85)
}
