% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_poc_nonadap.R
\name{single_poc}
\alias{single_poc}
\title{Single Arm Calculator (Non-Adaptive)}
\usage{
single_poc(soc = 0.25, exp = 0.4, alpha = 0.05, pwr = 0.8)
}
\description{
This function to calculate n given a desired alpha and power in a non-adaptive setting. By default considers the 25% vs 40% case for alpha = 0.05 and power = 0.80
}
